command-service/requirements.txt
fastapi==0.103.0
uvicorn==0.20.0
dataclasses-json==0.5.7
urllib3==2.0.7
pyyaml==6.0.1
backoff==2.2.1
